+++ R/rdd.R
@@ -170,6 +170,166 @@ parse_rdd_file <- function (path, mdd = FALSE) {
 }
 # }}}
 
+#' Format RddFile Object to Standard Input for `Idf$load()` Method
+#'
+#' `rdd_to_load()` and `mdd_to_load()` takes a `RddFile` and `MddFile` object
+#' respectively and format it into a [data.table][data.table::data.table()] in
+#' acceptable format for `$load()` method in [Idf] class.
+#'
+#' @param rdd,mdd A `RddFile` object created using [read_rdd()] and a `MddFile`
+#' object created using [read_mdd()], respectively.
+#' @param key_value Key value name for **all** variables. If not specified and
+#' the `key_value` column in the input `RddFile` object will be used. If
+#' `key_value` column does not exist, `"*"` are used for all variables.
+#' @param reporting_frequency Variable value reporting frequency for **all**
+#' variables. If not specified and the `reporting_freqency` column in the input
+#' `RddFile` object will be used. If `reporting_freqency` column does not exist,
+#' `"Timestep"` are used for all variables. All possible values: `"Detailed"`,
+#' `"Timestep"`, `"Hourly"`, `"Daily"`, `"Monthly"`, `"RunPeriod"`,
+#' `"Environment"`, and `"Annual"`.
+#' @param class Class name for meter output. All possible values:
+#' `"Output:Meter"`, `"Output:Meter:MeterFileOnly"`,
+#' `"Output:Meter:Cumulative"`, and `"Output:Meter:Cumulative:MeterFileOnly"`.
+#' Default: `"Output:Meter"`.
+#' @return
+#' A [data.table][data.table::data.table()] with 5 columns with an additional
+#' attribute named `eplus_version` extracted from the original `RddFile` and
+#' `MddFile`:
+#'
+#' * `id`: Integer type. Used to distinguish each object definition.
+#' * `class`: Character type. Class names, e.g. `Output:Variable` and
+#'   `Output:Meter`.
+#' * `index`: Integer type. Field indices.
+#' * `field`: Character type. Field names.
+#' * `value`: Character type. The value of each field to be added.
+#'
+#' @rdname rdd_to_load
+#' @export
+#' @examples
+#' \dontrun{
+#' idf <- read_idf(system.file("extdata/1ZoneUncontrolled.idf", package = "eplusr"))
+#' job <- idf$run(dir = tempdir())
+#' rdd_to_load(job$read_rdd())
+#' mdd_to_load(job$read_mdd())
+#'
+#' # using $load() to add output objects
+#' idf$load(mdd_to_load(job$read_mdd()))
+#' }
+#' @export
+# rdd_to_load {{{
+rdd_to_load <- function (rdd, key_value, reporting_frequency) {
+    assert(is_rdd(rdd))
+
+    # copy the original
+    rdd <- copy(rdd)
+    ver <- attr(rdd, "eplus_version")
+
+    # update index
+    rdd[, index := .I]
+
+    set(rdd, NULL, "class", "Output:Variable")
+
+    if (!missing(key_value)) {
+        assert(is_string(key_value))
+        set(rdd, NULL, "key_value", key_value)
+    } else if (!has_name(rdd, "key_value")) {
+        set(rdd, NULL, "key_value", "*")
+    } else {
+        set(rdd, NULL, "key_value", as.character(rdd$key_value))
+    }
+
+    if (!missing(reporting_frequency)) {
+        rep_freq <- validate_report_freq(reporting_frequency)
+        set(mdd, NULL, "reporting_frequency", rep_freq)
+    } else if (!has_name(rdd, "reporting_frequency")) {
+        set(rdd, NULL, "reporting_frequency", "Timestep")
+    } else {
+        set(rdd, NULL, "reporting_frequency", as.character(rdd$reporting_frequency))
+        validate_report_freq(rdd$reporting_frequency, scalar = FALSE)
+    }
+
+    setnames(rdd, c("index", "key_value", "variable", "reporting_frequency"),
+        c("id", "Key Value", "Variable Name", "Reporting Frequency")
+    )
+
+    rdd <- data.table::melt.data.table(rdd, id.vars = c("id", "class"),
+        measure.vars = c("Key Value", "Variable Name", "Reporting Frequency"),
+        variable.name = "field", variable.factor = FALSE
+    )
+    set(rdd, NULL, "index", rowidv(rdd, "id"))
+    setorderv(rdd, c("id", "index"))
+    setcolorder(rdd, c("id", "class", "index", "field", "value"))
+    setattr(rdd, "eplus_version", ver)
+    rdd[]
+}
+# }}}
+
+#' @rdname rdd_to_load
+#' @export
+# mdd_to_load {{{
+mdd_to_load <- function (mdd, reporting_frequency, class = c("Output:Meter",
+                                                             "Output:Meter:MeterFileOnly",
+                                                             "Output:Meter:Cumulative",
+                                                             "Output:Meter:Cumulative:MeterFileOnly")) {
+    assert(is_mdd(mdd))
+    ver <- attr(mdd, "eplus_version")
+    class <- match.arg(class)
+
+    # copy the original
+    mdd <- copy(mdd)
+
+    # update index
+    mdd[, index := .I]
+
+    set(mdd, NULL, "class", class)
+
+    if (!missing(reporting_frequency)) {
+        rep_freq <- validate_report_freq(reporting_frequency)
+        set(mdd, NULL, "reporting_frequency", rep_freq)
+    } else if (!has_name(mdd, "reporting_frequency")) {
+        set(mdd, NULL, "reporting_frequency", "Timestep")
+    } else {
+        set(mdd, NULL, "reporting_frequency", as.character(mdd$reporting_frequency))
+        validate_report_freq(mdd$reporting_frequency, scalar = FALSE)
+    }
+
+    setnames(mdd, c("index", "variable", "reporting_frequency"),
+        c("id", "Key Name", "Reporting Frequency")
+    )
+
+    mdd <- data.table::melt.data.table(mdd, id.vars = c("id", "class"),
+        measure.vars = c("Key Name", "Reporting Frequency"),
+        variable.name = "field", variable.factor = FALSE
+    )
+    set(mdd, NULL, "index", rowidv(mdd, "id"))
+    setorderv(mdd, c("id", "index"))
+    setcolorder(mdd, c("id", "class", "index", "field", "value"))
+    setattr(mdd, "eplus_version", ver)
+    mdd[]
+}
+# }}}
+
+# validate_report_freq {{{
+validate_report_freq <- function (reporting_frequency, scalar = TRUE) {
+    if (scalar) assert(is_string(reporting_frequency))
+
+    all_freq <- c("Detailed", "Timestep", "Hourly", "Daily", "Monthly",
+          "RunPeriod", "Environment", "Annual")
+
+    freq <- match_in_vec(reporting_frequency, all_freq, label = TRUE)
+
+    assert(no_na(freq),
+        msg = paste0("Invalid reporting frequency found: ",
+            collapse(reporting_frequency[is.na(freq)]), ". All possible values: ",
+            collapse(all_freq), "."
+        ),
+        err_type = "error_invalid_reporting_frequency"
+    )
+
+    freq
+}
+# }}}
+
 #' @export
 # print.RddFile {{{
 print.RddFile <- function (x, ...) {